extract out build_views to use with varbin
